Webhook clean files when process done

+++ tools.py
@@ -43,17 +43,21 @@ def get_FormsPath():
   create_directory(PA.FulcrumPath+"forms/")
   return PA.FulcrumPath+"forms/"
 def get_WebhookFormsPath():
-  create_directory(PA.FulcrumWebhook+"forms/")
+  #create_directory(PA.FulcrumWebhook+"forms/")
   return PA.FulcrumWebhook+"forms/"
+def get_WebhookRecordsPath():
+  return PA.FulcrumWebhook+"records/"
+
 
 # try to delete a file
 def delete_a_file(fname,logName="main"):
-  try:
-    os.remove(fname)
-    return True
-  except Exception, e:
-    LO.l_err(e,logName)
-    return False
+  if os.path.isfile(fname):
+    try:
+      os.remove(fname)
+      return True
+    except Exception, e:
+      LO.l_err(e,logName)
+  return False
 
 # Test if a file is available and not empty
 def file_is_here(fname,logName="main"):
